iommu: bcm2712-iommu: Define MMU_PAGE_SIZE separate from PAGE_SIZE
BCM2712 IOMMU uses a page size of 4K bytes but no longer assumes
Linux PAGE_SIZE is equal to this (it might be larger, e.g. 16K).
The IOMMU now advertise 16K as one of its supported page sizes.
Minor fixes and tidying.


(cherry picked from commit 4289a89a8a15d68b71feab969c032b20bcde618c pi5_61)

+++ drivers/iommu/bcm2712-iommu.h
@@ -29,8 +29,8 @@ struct bcm2712_iommu {
 	spinlock_t hw_lock;   /* to protect HW registers */
 	void __iomem *reg_base;
 	u64 dma_iova_offset; /* Hack for IOMMU attached to PCIe RC */
-	u8 bigpage_mask;
-	u8 superpage_mask;
+	u32 bigpage_mask;
+	u32 superpage_mask;
 	unsigned int nmapped_pages;
 	bool dirty; /* true when tables are oriented towards CPU */
 };
